Extend login_to_service on filter to reuse config.
In Rails, when writing an autologin function (ie send credentials to
rubyCAS-server instead of redirecting user), we want to reuse the existing
config, not have to set up a brand new Client each time and repeat the
config.

So I've extended the filter to add a login_to_service method.

Putting this on the filter is probably the wrong place to put it... but at
least we already know it's there, along with the known config.

I've also included some brief rdoc on how to use this function to use a
login-from-a-form.